move extensions to core_ext
